Add PN mode sheet & fix unsubscribing
