+++ Signal/translations/en.lproj/Localizable.strings
@@ -2818,3 +2818,4 @@
 "Use APNs" = "Use APNs";
 "Recommended" = "Recommended";
 "Notification Strategy" = "Notification Strategy";
+"Session now features two ways to handle push notifications. Make sure to read the descriptions carefully before you choose." = "Session now features two ways to handle push notifications. Make sure to read the descriptions carefully before you choose.";
