+++ SignalServiceKit/src/Loki/API/LokiPushNotificationManager.swift
@@ -17,8 +17,7 @@ public final class LokiPushNotificationManager : NSObject {
     // MARK: Registration
     /// Registers the user for silent push notifications (that then trigger the app
     /// into fetching messages). Only the user's device token is needed for this.
-    @objc(registerWithToken:)
-    static func register(with token: Data) {
+    static func register(with token: Data) -> Promise<Void> {
         let hexEncodedToken = token.toHexString()
         let userDefaults = UserDefaults.standard
         let oldToken = userDefaults[.deviceToken]
@@ -26,16 +25,18 @@ public final class LokiPushNotificationManager : NSObject {
         let isUsingFullAPNs = userDefaults[.isUsingFullAPNs]
         let now = Date().timeIntervalSince1970
         guard hexEncodedToken != oldToken || now - lastUploadTime < tokenExpirationInterval else {
-            return print("[Loki] Device token hasn't changed; no need to re-upload.")
+            print("[Loki] Device token hasn't changed; no need to re-upload.")
+            return Promise<Void> { $0.fulfill(()) }
         }
         guard !isUsingFullAPNs else {
-            return print("[Loki] Using full APNs; ignoring call to register(with:).")
+            print("[Loki] Using full APNs; ignoring call to register(with:).")
+            return Promise<Void> { $0.fulfill(()) }
         }
         let parameters = [ "token" : hexEncodedToken ]
         let url = URL(string: server + "register")!
         let request = TSRequest(url: url, method: "POST", parameters: parameters)
         request.allHTTPHeaderFields = [ "Content-Type" : "application/json" ]
-        TSNetworkManager.shared().makeRequest(request, success: { _, response in
+        let promise = TSNetworkManager.shared().makePromise(request: request).map { _, response in
             guard let json = response as? JSON else {
                 return print("[Loki] Couldn't register device token.")
             }
@@ -45,9 +46,19 @@ public final class LokiPushNotificationManager : NSObject {
             userDefaults[.deviceToken] = hexEncodedToken
             userDefaults[.lastDeviceTokenUpload] = now
             userDefaults[.isUsingFullAPNs] = false
-        }, failure: { _, error in
+            return
+        }
+        promise.catch { error in
             print("[Loki] Couldn't register device token.")
-        })
+        }
+        return promise
+    }
+
+    /// Registers the user for silent push notifications (that then trigger the app
+    /// into fetching messages). Only the user's device token is needed for this.
+    @objc(registerWithToken:)
+    static func objc_register(with token: Data) -> AnyPromise {
+        return AnyPromise.from(register(with: token))
     }
 
     /// Registers the user for normal push notifications. Requires the user's device
@@ -78,6 +89,8 @@ public final class LokiPushNotificationManager : NSObject {
         return promise
     }
 
+    /// Registers the user for normal push notifications. Requires the user's device
+    /// token and their Session ID.
     @objc(registerWithToken:hexEncodedPublicKey:)
     static func objc_register(with token: Data, hexEncodedPublicKey: String) -> AnyPromise {
         return AnyPromise.from(register(with: token, hexEncodedPublicKey: hexEncodedPublicKey))
