+++ SignalServiceKit/src/Loki/Utilities/LKUserDefaults.swift
@@ -5,6 +5,7 @@ public enum LKUserDefaults {
     public enum Bool : Swift.String {
         case hasLaunchedOnce
         case hasSeenOpenGroupSuggestionSheet
+        case hasSeenPNModeSheet
         case hasViewedSeed
         /// Whether the device was unlinked as a slave device (used to notify the user on the landing screen).
         case wasUnlinked
